refactor: combine DigiKey + Mouser API servers into one (issue #26)
Folds api/digikey/server/ and api/mouser/server/ into a single
api/server/ FastAPI app. One process on port 8000, one .env with all
three credentials (DIGIKEY_CLIENT_ID, DIGIKEY_CLIENT_SECRET,
MOUSER_API_KEY), one CORS middleware block, one /health, one venv to
manage.

Routing is namespaced by distributor:
  - GET /digikey/pricing?manufacturer_part_number=<MPN>
  - GET /mouser/pricing?manufacturer_part_number=<MPN>

Same ?manufacturer_part_number= param, same response shape as before
the refactor.

digikey_client.py and mouser_client.py move unchanged via git mv —
distributor-specific logic still lives in dedicated client modules; the
new code is just the routing layer (api/server/main.py with two
APIRouter mounts).

Frontends updated:
  - web/mouser-search.html port 8001 → 8000
  - both pages fetch /digikey/pricing or /mouser/pricing (was bare
    /pricing)
  - error-message README pointers updated to api/server/README.md

web/TEST-PLAN.md §§ 6c, 7c smoke-test commands updated. § 7c.5
reframed from "two backends side by side" to "one backend serving
both at different path prefixes."

Smoke-tested:
  - GET /health → 200 {"status":"ok"}
  - GET /mouser/pricing → 200 live data (Qty 4000 → $0.224 USD for
    NE555P)
  - GET /digikey/pricing → 502 clean "missing credentials" error
    (no DigiKey creds in this .env; routing layer verified)
