--- api/mouser/server/.env.example
@@ -1,1 +1,0 @@
-MOUSER_API_KEY=
