--- api/mouser/server/requirements.txt
@@ -1,4 +1,0 @@
-fastapi>=0.110
-uvicorn[standard]>=0.27
-httpx>=0.27
-python-dotenv>=1.0
