+++ api/server/.env.example
@@ -1,2 +1,3 @@
 DIGIKEY_CLIENT_ID=
 DIGIKEY_CLIENT_SECRET=
+MOUSER_API_KEY=
