+++ web/TEST-PLAN.md
@@ -117,7 +117,7 @@ This feature was originally a static day-of-year-rotating array (issue #2) and w
 
 ## 6. Product search (issue #20)
 
-The `web/digikey-search.html` page is a static frontend that calls a **local** Python FastAPI backend at `http://localhost:8000` (`api/digikey/server/`). The live deployed page exists but is non-functional unless the user has the backend running with their own DigiKey credentials.
+The `web/digikey-search.html` page is a static frontend that calls a **local** Python FastAPI backend at `http://localhost:8000/digikey/...` (`api/server/`). The live deployed page exists but is non-functional unless the user has the combined backend running with their own DigiKey credentials. Since #26, the DigiKey and Mouser backends share one process — see § 7 for the Mouser-side smoke tests against the same server.
 
 ### 6a. Homepage link
 
@@ -130,16 +130,16 @@ The `web/digikey-search.html` page is a static frontend that calls a **local** P
 
 | ID | Steps | Expected |
 |---|---|---|
-| 6b.1 | On the live `/digikey-search.html` with no local backend running, type `STM32F407VGT6` and submit. | After a brief "Loading…", an error message appears: "Backend unreachable at `http://localhost:8000` — start the local server (see `api/digikey/server/README.md`)." Form re-enables; no console crash. |
+| 6b.1 | On the live `/digikey-search.html` with no local backend running, type `STM32F407VGT6` and submit. | After a brief "Loading…", an error message appears: "Backend unreachable at `http://localhost:8000` — start the local server (see `api/server/README.md`)." Form re-enables; no console crash. |
 
 ### 6c. Local backend smoke (developer-only)
 
-Prerequisite: backend up per `api/digikey/server/README.md` with real `DIGIKEY_CLIENT_ID` / `DIGIKEY_CLIENT_SECRET`.
+Prerequisite: combined backend up per `api/server/README.md` with real `DIGIKEY_CLIENT_ID` / `DIGIKEY_CLIENT_SECRET` (and `MOUSER_API_KEY` if you'll also run § 7c).
 
 | ID | Steps | Expected |
 |---|---|---|
 | 6c.1 | `curl http://localhost:8000/health` | Returns `{"status":"ok"}`. |
-| 6c.2 | `curl 'http://localhost:8000/pricing?manufacturer_part_number=STM32F407VGT6'` | HTTP 200 with JSON containing `manufacturer_part_number`, `digikey_part_number`, `currency: "USD"`, non-empty `tiers` array (sorted ascending by `quantity`), and a `unit_price` / `tier_quantity` pair matching `tiers[-2]` (or `tiers[0]` if fewer than 2 entries). |
+| 6c.2 | `curl 'http://localhost:8000/digikey/pricing?manufacturer_part_number=STM32F407VGT6'` | HTTP 200 with JSON containing `manufacturer_part_number`, `digikey_part_number`, `currency: "USD"`, non-empty `tiers` array (sorted ascending by `quantity`), and a `unit_price` / `tier_quantity` pair matching `tiers[-2]` (or `tiers[0]` if fewer than 2 entries). |
 | 6c.3 | Serve `web/` locally (`cd web && python -m http.server 8080`), open `http://localhost:8080/digikey-search.html`, type a known MPN, submit. | Headline renders as `Qty <N> → $<P> USD / unit` with both numbers in the same large font weight/size. A muted line below shows `<MPN>  ·  DK: <DigiKey-PN>`. No tier table appears. |
 | 6c.4 | Submit a bogus MPN like `NOTAPART_xyz`. | After "Loading…", an error message renders (DigiKey 404 surfaced as a 502 from the proxy with a "Part not found" detail), not the headline. |
 
@@ -151,7 +151,7 @@ Prerequisite: backend up per `api/digikey/server/README.md` with real `DIGIKEY_C
 
 ## 7. Mouser product search (issue #24)
 
-The `web/mouser-search.html` page is a static frontend that calls a **local** Python FastAPI backend at `http://localhost:8001` (`api/mouser/server/`). Same shape as § 6 but for Mouser instead of DigiKey, and on a different port so both backends can run side by side.
+The `web/mouser-search.html` page is a static frontend that calls a **local** Python FastAPI backend at `http://localhost:8000/mouser/...` (`api/server/`). Same shape as § 6 but for Mouser; since #26 both distributors share the single backend on port 8000, so § 6c and § 7c hit the same `uvicorn` process at different path prefixes.
 
 ### 7a. Homepage link
 
@@ -164,19 +164,19 @@ The `web/mouser-search.html` page is a static frontend that calls a **local** Py
 
 | ID | Steps | Expected |
 |---|---|---|
-| 7b.1 | On the live `/mouser-search.html` with no local backend running, type `NE555P` and submit. | After a brief "Loading…", an error message appears: "Backend unreachable at `http://localhost:8001` — start the local server (see `api/mouser/server/README.md`)." Form re-enables; no console crash. |
+| 7b.1 | On the live `/mouser-search.html` with no local backend running, type `NE555P` and submit. | After a brief "Loading…", an error message appears: "Backend unreachable at `http://localhost:8000` — start the local server (see `api/server/README.md`)." Form re-enables; no console crash. |
 
 ### 7c. Local backend smoke (developer-only)
 
-Prerequisite: backend up per `api/mouser/server/README.md` with a real `MOUSER_API_KEY`.
+Prerequisite: combined backend up per `api/server/README.md` with a real `MOUSER_API_KEY` (and DigiKey credentials if you'll also run § 6c against the same process).
 
 | ID | Steps | Expected |
 |---|---|---|
-| 7c.1 | `curl http://localhost:8001/health` | Returns `{"status":"ok"}`. |
-| 7c.2 | `curl 'http://localhost:8001/pricing?manufacturer_part_number=NE555P'` | HTTP 200 with JSON containing `manufacturer_part_number`, `mouser_part_number`, `currency`, non-empty `tiers` array (sorted ascending by `quantity`), and a `unit_price` / `tier_quantity` pair matching `tiers[-2]` (or `tiers[0]` if fewer than 2 entries). |
+| 7c.1 | `curl http://localhost:8000/health` | Returns `{"status":"ok"}`. (Same `/health` endpoint covers both distributors.) |
+| 7c.2 | `curl 'http://localhost:8000/mouser/pricing?manufacturer_part_number=NE555P'` | HTTP 200 with JSON containing `manufacturer_part_number`, `mouser_part_number`, `currency`, non-empty `tiers` array (sorted ascending by `quantity`), and a `unit_price` / `tier_quantity` pair matching `tiers[-2]` (or `tiers[0]` if fewer than 2 entries). |
 | 7c.3 | Serve `web/` locally (`cd web && python3 -m http.server 8080`), open `http://localhost:8080/mouser-search.html`, type a known MPN, submit. | Headline renders as `Qty <N> → $<P> USD / unit` with both numbers in the same large font weight/size. A muted line below shows `<MPN>  ·  Mouser: <Mouser-PN>`. No tier table appears. |
 | 7c.4 | Submit a bogus MPN like `NOTAPART_xyz`. | After "Loading…", an error message renders ("Part not found" surfaced as a 502 from the proxy), not the headline. |
-| 7c.5 | Run both backends side by side: `uvicorn main:app --port 8000` from `api/digikey/server` and `--port 8001` from `api/mouser/server`. Open both `digikey-search.html` and `mouser-search.html` in two tabs. | Each page only talks to its own backend; searches in one tab don't affect the other. |
+| 7c.5 | With one `uvicorn main:app --port 8000` from `api/server/`, open both `digikey-search.html` and `mouser-search.html` in two tabs simultaneously. Search a known MPN on each. | Each page hits its own path prefix (`/digikey/pricing` vs `/mouser/pricing`) on the same backend process. Both return their respective headlines. Network tab shows both responses coming from `localhost:8000`. |
 
 ### 7d. Defense-in-depth
