+++ web/digikey-search.html
@@ -193,7 +193,7 @@
       submitBtn.disabled = true;
       setMessage('Loading…', 'muted');
       try {
-        const url = BACKEND_URL + '/pricing?manufacturer_part_number=' + encodeURIComponent(mpn);
+        const url = BACKEND_URL + '/digikey/pricing?manufacturer_part_number=' + encodeURIComponent(mpn);
         const res = await fetch(url);
         if (!res.ok) {
           let detail = 'HTTP ' + res.status;
@@ -208,7 +208,7 @@
       } catch (err) {
         const msg = err && err.message ? err.message : String(err);
         if (msg.includes('Failed to fetch') || msg.includes('NetworkError')) {
-          setMessage('Backend unreachable at ' + BACKEND_URL + ' — start the local server (see api/digikey/server/README.md).', 'error');
+          setMessage('Backend unreachable at ' + BACKEND_URL + ' — start the local server (see api/server/README.md).', 'error');
         } else {
           setMessage(msg, 'error');
         }
