+++ web/mouser-search.html
@@ -125,7 +125,7 @@
   <div id="result" aria-live="polite"></div>
 
   <script>
-    const BACKEND_URL = "http://localhost:8001";
+    const BACKEND_URL = "http://localhost:8000";
 
     const form = document.getElementById('search-form');
     const submitBtn = document.getElementById('submit-btn');
@@ -183,7 +183,7 @@
       submitBtn.disabled = true;
       setMessage('Loading…', 'muted');
       try {
-        const url = BACKEND_URL + '/pricing?manufacturer_part_number=' + encodeURIComponent(mpn);
+        const url = BACKEND_URL + '/mouser/pricing?manufacturer_part_number=' + encodeURIComponent(mpn);
         const res = await fetch(url);
         if (!res.ok) {
           let detail = 'HTTP ' + res.status;
@@ -198,7 +198,7 @@
       } catch (err) {
         const msg = err && err.message ? err.message : String(err);
         if (msg.includes('Failed to fetch') || msg.includes('NetworkError')) {
-          setMessage('Backend unreachable at ' + BACKEND_URL + ' — start the local server (see api/mouser/server/README.md).', 'error');
+          setMessage('Backend unreachable at ' + BACKEND_URL + ' — start the local server (see api/server/README.md).', 'error');
         } else {
           setMessage(msg, 'error');
         }
